agents/ai-to-founder-voice-agent/requirements.txt
google-adk
google-cloud-firestore
google-cloud-storage
google-cloud-secret-manager
google-genai
python-dotenv
google-auth
PyMuPDF
requests
beautifulsoup4
tavily-python
aiohttp
fastapi
uvicorn[standard]